requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.57.0
anthropic==0.39.0
google-generativeai==0.8.2
httpx==0.27.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.6.0

